- polished code organization, tweaked movement + rotation
Should be good for playtesting!
