+++ unity-project/mini-game-collection/Assets/2024/Team10/Prefabs/2024-team10-prefab-redship.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fc046d4fe599363429e492e134ed6089
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
