--- unity-project/mini-game-collection/Assets/2024/Team10/Scripts/NewBehaviourScript.cs
@@ -1,24 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-
-namespace MiniGameCollection.Games2024.Team10
-{
-    // script code
-
-    public class NewBehaviourScript : MonoBehaviour
-    {
-        // Start is called before the first frame update
-        void Start()
-        {
-
-        }
-
-        // Update is called once per frame
-        void Update()
-        {
-
-        }
-    }
-}
