--- unity-project/mini-game-collection/Assets/2024/Team10/Scripts/middleDivider.cs
@@ -1,9 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class middleDivider : MonoBehaviour
-{
-    
-
-}
