--- unity-project/mini-game-collection/Assets/2024/Team10/Scripts/middleDivider.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8286cc755f668d4409201b28d3b1b3ed
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
