+++ unity-project/mini-game-collection/Assets/2024/Team10/Scripts/p2-movement.cs
@@ -1,14 +1,14 @@
 using System.Collections;
 using System.Collections.Generic;
+using Unity.Mathematics;
 using UnityEngine;
-
 
 namespace MiniGameCollection.Games2024.Team10
 {
     // script code
 
-public class p1movement : MonoBehaviour
-{
+    public class p2movement : MonoBehaviour
+    {
         private Rigidbody rigidBody;
         public float HandleSpeed = 14f;
         private Vector3 maxSpeed;
@@ -19,14 +19,12 @@ public class p1movement : MonoBehaviour
         public float dashDuration = 0f;
         public float dashUses = 3f;
         public Renderer playerRender;
-        public Color originalGreen;
+        public Color originalRed;
+        float xForceAdjustments = 25f;
+        float zForceAdjustments = 25f;
 
-        //float moveX = ArcadeInput.Player1.AxisX;
-        //bool moveDown = false;
-        //bool moveLeft = false;
-        //bool moveRight = false;
-        //bool beamButton = false;
-        //bool dashButton = false;
+        //Rotation
+        Vector3 movementDirect = Vector3.zero;
 
 
         void Start()
@@ -35,53 +33,77 @@ public class p1movement : MonoBehaviour
             maxSpeed = new Vector3(HandleSpeed, 0, HandleSpeed);
             rigidBody.angularDrag = angleDrag;
             //playerRender = GetComponent<Renderer>();
-            originalGreen = playerRender.material.color;
+            originalRed = playerRender.material.color;
         }
 
         // Update is called once per frame
         void Update()
         {
-            float xForceAdjustments = 25f;
-            float zForceAdjustments = 25f;
-            maxSpeed = new Vector3(HandleSpeed, 0, HandleSpeed);
-            if (Input.GetKey(KeyCode.A))
+            PlayerMovement();
+            PlayerDash();
+            
+
+            
+
+        }
+
+
+        public void PlayerMovement()
+        {
+            if (ArcadeInput.Player2.Left.Pressed || ArcadeInput.Player2.Left.Down)
+            {
                 if (Mathf.Abs(rigidBody.velocity.x) < maxSpeed.x)
                     rigidBody.AddForce(-xForceAdjustments * dashMultiplyer, 0, 0, ForceMode.Acceleration);
 
-            if (Input.GetKey(KeyCode.D))
+                movementDirect.z += 1f;
+            }
+
+            if (ArcadeInput.Player2.Right.Pressed || ArcadeInput.Player2.Right.Down)
+            {
                 if (rigidBody.velocity.x < maxSpeed.x)
                     rigidBody.AddForce(xForceAdjustments * dashMultiplyer, 0, 0, ForceMode.Acceleration);
+                movementDirect.z -= 1f;
+            }
 
-            if (Input.GetKey(KeyCode.S))
+            if (ArcadeInput.Player2.Down.Pressed || ArcadeInput.Player2.Right.Down)
+            {
                 if (Mathf.Abs(rigidBody.velocity.z) < maxSpeed.z)
                     rigidBody.AddForce(0, 0, -zForceAdjustments * dashMultiplyer, ForceMode.Acceleration);
 
-            if (Input.GetKey(KeyCode.W))
+                movementDirect.x -= 1f;
+            }
+            if (ArcadeInput.Player2.Up.Pressed || ArcadeInput.Player2.Right.Down)
+            {
                 if (rigidBody.velocity.z < maxSpeed.z)
                     rigidBody.AddForce(0, 0, zForceAdjustments * dashMultiplyer, ForceMode.Acceleration);
+                movementDirect.x += 1f;
+            }
 
 
-            //if (Input.GetKey(KeyCode.Q))
-            //{
-                
-            //    rigidBody.AddTorque(Vector3.up * -rotationSpeed * Time.deltaTime, ForceMode.VelocityChange);
-            //}
+            if (movementDirect != Vector3.zero)
+            {
+                movementDirect.Normalize();
+                movementDirect.y = 0f;
 
-            //if (Input.GetKey(KeyCode.E))
-            //{
-                
-            //    rigidBody.AddTorque(Vector3.up * rotationSpeed * Time.deltaTime, ForceMode.VelocityChange);
-            //}
+                Quaternion targetRotation = Quaternion.LookRotation(movementDirect);
 
-            if (Input.GetKey(KeyCode.E))
+                float rotationSpeed = 90f;
+                transform.rotation = Quaternion.RotateTowards(transform.rotation, targetRotation, rotationSpeed * Time.deltaTime);
+
+            }
+        }
+
+        public void PlayerDash()
+        {
+            if (ArcadeInput.Player2.Action2.Pressed || Input.GetKeyDown(KeyCode.Comma))
             {
                 if (dashUses > 0f && dashDuration < 0f)
                 {
                     dashUses--;
-                    dashMultiplyer = dashSpeed;                   
+                    dashMultiplyer = dashSpeed;
                     dashDuration = 5f;
-                    playerRender.material.color = Color.blue;
-                    
+                    playerRender.material.color = Color.magenta;
+
                 }
             }
 
@@ -92,11 +114,9 @@ public class p1movement : MonoBehaviour
             if (dashDuration < 0f)
             {
                 dashMultiplyer = 1f;
-                playerRender.material.color = originalGreen;
-                
+                playerRender.material.color = originalRed;
+
             }
-
         }
-
     }
 }
