+++ unity-project/mini-game-collection/Assets/2024/Team10/Scripts/red-goal.cs.meta
